A few scaling fixes for X11 mode...
- getscaling() incorrectly used an int intermediate variable
  This would force scaling to an integer value when read.
- Only parse ConfigureNotify events describing a change to the SyncTERM window
  It appears that XWayland on ChromeOS was occasionally sending a
  ConfigureNotify event describing some other window to SyncTERM,
  and that window had a 1x1 size, which would cause SyncTERM to
  assert minimum size again (usually 640x480), which would set
  scaling to 1.0.

+++ src/conio/x_cio.c
@@ -554,7 +554,7 @@ void x_setscaling(double newval)
 
 double x_getscaling(void)
 {
-	int ret;
+	double ret;
 
 	pthread_mutex_lock(&vstatlock);
 	ret = vstat.scaling;
